requirements.txt
PyYAML==5.4.1
numpy==1.19.5
scikit-learn==0.24.2
torch==1.8.1+cu111
transformers==4.6.0
pytorch-lightning==1.3.1
pandas==1.2.4
tqdm==4.49.0
accelerate==0.3.0
datasets==1.7.0
pyro_ppl == 1.6.0